moving copy-igvmfile to test.rs
