+++ Guide/src/dev_guide/dev_tools/vmgstool.md
@@ -67,9 +67,9 @@ variables without parsers:
 
 Additionally, the VmgsTool contains a tool to read the IGVMfile from a DLL (passed in as a data file)
 and write it to VMGS FileId 8 (GUEST_FIRMWARE). To do this pass one of five resource codes
-(nonconfidential, snp, tdx, snp_no_hcl, tdx_no_hcl) into the cmdline tool:
+(NONCONFIDENTIAL, SNP, TDX, SNP_NO_HCL, TDX_NO_HCL) into the cmdline tool:
 
-`vmgstool.exe copy-igvmfile --filepath <vmgs file path> --keypath <key file path> --datapath <dll path> --resource-code <code>`
+`vmgstool.exe test copy-igvmfile --filepath <vmgs file path> --datapath <dll path> --resource-code <code>`
 
 ### Delete Boot Variables to Recover a VM that Fails to Boot
